Dodati modali za sign in i sign up, sign up povezan sa backom

+++ uber-clone/src/main/java/com/backend/uberclone/repository/RideRepository.java
@@ -4,13 +4,14 @@ import com.backend.uberclone.model.Ride;
 import com.backend.uberclone.model.RideStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Collection;
 import java.util.List;
 
 public interface RideRepository extends JpaRepository<Ride,Integer> {
-    Ride findByIdAndStatusIn(Long id, Collection<RideStatus> status);
-    Ride findByIdAndStatus(Long id, RideStatus status);
-    Ride findByIdAndStatusAndDriverId(Long id, RideStatus status, Long driver_id);
-    Ride findByIdAndStatusInAndDriverId(Long id, Collection<RideStatus> status, Long driverId);
+    Ride findByIdAndStatusIn(Integer id, Collection<RideStatus> status);
+    Ride findByIdAndStatus(Integer id, RideStatus status);
+    Ride findByIdAndStatusAndDriverId(Long id, RideStatus status, Integer driver_id);
+    Ride findByIdAndStatusInAndDriverId(Long id, Collection<RideStatus> status, Integer driverId);
 
     Ride findOneById(int rideId);
 }
